Reduce non-functional changes
